Refactoring Pattern
* Shoes3/shoes3#333
