+++ shoes/types/textblock.h
@@ -43,4 +43,33 @@ void shoes_textblock_send_release(VALUE self, int button, int x, int y);
 
 // canvas
 
+#define GET_STYLE(name) \
+  attr = NULL; \
+  str = Qnil; \
+  if (!NIL_P(oattr)) str = rb_hash_aref(oattr, ID2SYM(s_##name)); \
+  if (!NIL_P(hsh) && NIL_P(str)) str = rb_hash_aref(hsh, ID2SYM(s_##name))
+     
+#define APPLY_ATTR() \
+  if (attr != NULL) { \
+    attr->start_index = start_index; \
+    attr->end_index = end_index; \
+    pango_attr_list_insert_before(block->pattr, attr); \
+    attr = NULL; \
+  }
+
+#define APPLY_STYLE_COLOR(name, func) \
+  GET_STYLE(name); \
+  if (!NIL_P(str)) \
+  { \
+    if (TYPE(str) == T_STRING) \
+      str = shoes_color_parse(cColor, str); \
+    if (rb_obj_is_kind_of(str, cColor)) \
+    { \
+      shoes_color *color; \
+      Data_Get_Struct(str, shoes_color, color); \
+      attr = pango_attr_##func##_new(color->r * 255, color->g * 255, color-> b * 255); \
+    } \
+    APPLY_ATTR(); \
+  }
+
 #endif
